+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Remove useless console.log
 
+### Added
+- [PREVIEW] Add path completion for local links
+
 
 ## [1.3.17] - 20220609
 ### Changed